requirement.txt
# Install Dependencies
langchain==0.3.20
langchain-community==0.3.19
langchain-openai==0.3.8
python-docx==1.0.1
openai==1.65.5
pydantic 
PyYAML 
python-dotenv==1.0.0
faiss-cpu==1.10.0
pypdf==5.3.1
anthropic==0.49.0
markdown-pdf
flask==2.2.3
werkzeug==2.2.3
flask-session==0.4.0
# google-serper package not available, using langchain-community for search instead